backend/pyproject.toml
[project]
name = "template"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]==0.135.3",
    "pre-commit==4.5.1",
    "psycopg2-binary==2.9.11",
    "pytest==9.0.2",
    "pyyaml==6.0.3",
    "ruff==0.15.8",
    "ty==0.0.27",
    "uvicorn==0.42.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
